Added test for hourly employee payday. Moved from decimal to double with money and rates.

+++ SalaryRCM/Models/Affiliation.cs
@@ -4,6 +4,6 @@ namespace PayrollSystem.Models
 {
     public abstract class Affiliation
     {
-        public abstract decimal CalculatePay(DateTime paycheckDate);
+        public abstract double CalculatePay(DateTime paycheckDate);
     }
 }
