+++ SalaryRCM/Models/PaymentClassifications/PaymentClassification.cs
@@ -4,6 +4,6 @@ namespace PayrollSystem.Models.PaymentClassifications
 {
     public abstract class PaymentClassification
     {
-        public abstract decimal CalculatePay(DateTime paycheckDate);
+        public abstract double CalculatePay(DateTime paycheckDate);
     }
 }
